Updated and replaced all occurrences from population_functions to PopulationFunctions! Renamed population_functions.py to PopulationFunctions.py!
