+++ BackEndFlask/Functions/randAssign.py
@@ -1,5 +1,5 @@
 from Functions.customExceptions import *
-from Functions.test_files.population_functions import *
+from Functions.test_files.PopulationFunctions import *
 from models.course import *
 from models.user_course import *
 from models.team import *
