+++ BackEndFlask/Functions/test_files/test_randAssignTeams.py
@@ -1,5 +1,5 @@
 from Functions.customExceptions import *
-from population_functions import *
+from Functions.test_files.PopulationFunctions import *
 from models.team import *
 from models.user import *
 from models.team_user import *
